requirements.txt
tensorflow
numpy
pandas
opencv-python
pillow
gdown